Move events to AppSec specific namespace
The general use case is as of yet unclear. This will be AppSec specific
for now.
